# stage-pfe-front/tools/swagger/src/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger API documentation for Gestion de stock application",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger API Documentation for Gestion de stock application",
    "contact" : {
      "name" : "Ali Bouali",
      "url" : "https://www.udemy.com/user/bouali-ali/",
      "email" : "bouali.social@gmail.com"
    }
  },
  "basePath" : "http://localhost:8080",
  "paths" : {
...
    "/api/sessions/date-range" : {
      "get" : {
        "operationId" : "findByDateBetween",
        "parameters" : [ {
          "name" : "startDate",
          "in" : "query",
          "required" : true,
          "type" : "string",
          "format" : "date"
        }, {
          "name" : "endDate",
          "in" : "query",
          "required" : true,
          "type" : "string",
          "format" : "date"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SessionsDto"
              }
            }
          }
        }
      }
    },
    "/api/sessions/findByFormateurid/{formateurId}" : {
      "get" : {
        "operationId" : "findByFormateur",
        "parameters" : [ {
          "name" : "formateurId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SessionsDto"
              }
            }
          }
        }
      }
    },
    "/api/sessions/formation/{formationId}" : {
      "get" : {
        "operationId" : "findByFormation",
        "parameters" : [ {
          "name" : "formationId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SessionsDto"
              }
            }
          }
        }
      }
    },
...
    "/api/sessions/{id}" : {
      "get" : {
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SessionsDto"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/api/sessions/{sessionId}/available-places" : {
      "get" : {
        "operationId" : "getAvailablePlaces",
        "parameters" : [ {
          "name" : "sessionId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "integer",
              "format" : "int32"
            }
          }
        }
      }
    },
...
    "/api/sessions/{sessionId}/participants" : {
      "get" : {
        "operationId" : "getParticipants",
        "parameters" : [ {
          "name" : "sessionId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Utilisateurs"
              }
            }
          }
        }
      }
    },
...
    "/api/themes/{id}" : {
      "get" : {
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ThemesDto"
            }
          }
        }
      },
      "put" : {
        "operationId" : "updateTheme",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ThemesDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ThemesDto"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/save" : {
      "post" : {
        "operationId" : "save",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SallesDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SallesDto"
            }
          }
        }
      }
    },
    "/{id}" : {
      "get" : {
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SallesDto"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "SallesDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "capacite" : {
          "type" : "integer",
          "format" : "int32"
        },
        "equipement" : {
          "type" : "string"
        },
        "statut" : {
          "type" : "string"
        }
      }
    },
    "SessionsDto" : {
      "type" : "object",
      "properties" : {
        "sessionId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "formationId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "utilisateursIds" : {
          "type" : "array",
          "items" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "salleId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateDebut" : {
          "type" : "string",
          "format" : "date"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date"
        },
        "capacite" : {
          "type" : "integer",
          "format" : "int32"
        },
        "placesDisponibles" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "ThemesDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "formationsIds" : {
          "type" : "array",
          "items" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    },
    "Utilisateurs" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "photo" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "motDePasse" : {
          "type" : "string"
        },
        "etablissement" : {
          "type" : "string"
        },
        "cin" : {
          "type" : "string"
        },
        "dateNaissance" : {
          "type" : "string",
          "format" : "date"
        },
        "telephone" : {
          "type" : "string"
        },
        "role" : {
          "type" : "string",
          "enum" : [ "EXTERNE", "INTERNE", "ETUDIANT", "ADMIN", "EMPLOYEE" ]
        }
      }
    },
...
  }
}